tests/wpt/mozilla/tests/mozilla/bluetooth/requestDevice/name-empty-device-from-service-filter.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.empty_name_heart_rate);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [heart_rate.name]}]
    })
    .then(device => assert_equals(device.name, ''))
    .catch(error => assert_unreached('Promise should have resolved. ' + error));
}, 'An empty name device can be obtained by advertised service UUID.');
</script>
